Add path_sha to the Blob and Tree objects
